\name{otherworlds_continuous_palettes}
\alias{otherworlds_continuous_palettes}
\title{List Available Continuous Other Worlds Palettes}
\usage{
otherworlds_continuous_palettes()
}
\value{A named integer vector containing the number of colour anchors in each
continuous palette.}
\description{Lists the continuous palettes included in the package.}
\examples{
otherworlds_continuous_palettes()
}
