\name{otherworlds_palettes}
\alias{otherworlds_palettes}
\title{List Available Other Worlds Palettes}
\usage{otherworlds_palettes()}
\value{A named integer vector giving the maximum size of each palette.}
\description{Lists the palettes included in the package and their capacities.}
\examples{otherworlds_palettes()}
